market_intelligence/requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# requirements.txt — Market Intelligence Platform
# Pin major versions; let patch versions float for security fixes.
# ─────────────────────────────────────────────────────────────────────────────

# ── Django core ───────────────────────────────────────────────────────────────
Django>=4.2,<5.0
djangorestframework>=3.15,<4.0
django-environ>=0.11,<0.12
python-dotenv>=1.0,<2.0

# ── Database ──────────────────────────────────────────────────────────────────
psycopg2-binary>=2.9,<3.0

# ── Celery & task scheduling ──────────────────────────────────────────────────
celery[redis]>=5.3,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.6,<3.0
django-celery-results>=2.5,<3.0

# ── Cache (Redis ORM backend) ─────────────────────────────────────────────────
django-redis>=5.4,<6.0

# ── HTTP clients & scraping ────────────────────────────────────────────────────
requests>=2.31,<3.0
httpx>=0.27,<0.28                  # async-friendly alternative

# ── Reddit API ────────────────────────────────────────────────────────────────
praw>=7.7,<8.0

# ── NLP / Sentiment ───────────────────────────────────────────────────────────
textblob>=0.18,<0.19               # swap with VADER or HuggingFace in prod
vaderSentiment>=3.3,<4.0

# ── Data processing ────────────────────────────────────────────────────────────
pandas>=2.2,<3.0
numpy>=1.26,<2.0

# ── ML / Forecasting (install as needed) ──────────────────────────────────────
scikit-learn>=1.4,<2.0
xgboost>=2.0,<3.0
prophet>=1.1,<2.0                  # formerly fbprophet

# ── Serialization helpers ──────────────────────────────────────────────────────
marshmallow>=3.21,<4.0

# ── Production server ─────────────────────────────────────────────────────────
gunicorn>=21.2,<22.0
whitenoise>=6.6,<7.0               # static file serving

# ── Monitoring / tracing ──────────────────────────────────────────────────────
sentry-sdk[django]>=1.44,<2.0      # optional but recommended for prod

# ── Dev / testing ─────────────────────────────────────────────────────────────
pytest>=8.1,<9.0
pytest-django>=4.8,<5.0
factory-boy>=3.3,<4.0
coverage>=7.4,<8.0
